Refactor EnvFileWriter and associated files
Update EnvFileWriter so that it can be used to write any of the variables to the .env file not just the database connection variables. This means that it will be possible to easily and consistently add prompts for additional variables going forwards (e.g. APP_URL). Additionally, update ConfigureDatabase and SetInstalled flag so that they both use the updated CreateEnvFile. The .env file is now created by CreateEnvFile.